[US206] perfect 1:1

+++ emulator/opengl/host/libs/libOpenglRender/AndroVM.cpp
@@ -17,8 +17,8 @@ bool AndroVM_initOpenGLRenderer(int, int, int, OnPostFn, void*);
 void AndroVM_setCallbackRotation(void (* fn)(float));
 void AndroVM_repaintOpenGLDisplay();
 void AndroVM_setDPI(int);
-void AndroVM_setViewport(float width, float height);
-void AndroVM_scrollViewport(float x, float y);
+void AndroVM_setViewport(int width, int height);
+void AndroVM_scrollViewport(int x, int y);
 float AndroVM_getDisplayRotation();
 bool AndroVM_registerOGLCallback(OnPostFn, void*);
 void AndroVM_setLogo(char* logo, int width, int height);
@@ -116,12 +116,12 @@ void AndroVM_setDPI(int d) {
     setDPI(d);
 }
 
-void AndroVM_setViewport(float width, float height)
+void AndroVM_setViewport(int width, int height)
 {
     FrameBuffer::setViewport(width, height);
 }
 
-void AndroVM_scrollViewport(float x, float y)
+void AndroVM_scrollViewport(int x, int y)
 {
     FrameBuffer::scrollViewport(x, y);
 }
